Applying full set of internal changes to enable UWP as a target platform based on 4.11.2 release

+++ Engine/Source/Runtime/Sockets/Private/BSDIPv6Sockets/SocketsBSDIPv6.cpp
@@ -163,7 +163,9 @@ bool FSocketBSDIPv6::SendTo(const uint8* Data, int32 Count, int32& BytesSent, co
 	if(BytesSent == SOCKET_ERROR)
 	{
 		ESocketErrors SockError = SocketSubsystem->GetLastErrorCode();
-		UE_LOG(LogSockets, Log, TEXT("sendto error: (ESocketErrors:%d)"), SockError);
+		// @ATG_CHANGE : BEGIN strict enum fix
+		UE_LOG(LogSockets, Log, TEXT("sendto error: (ESocketErrors:%d)"), static_cast<int32>(SockError));
+		// @ATG_CHANGE : END
 	}
 //	NETWORK_PROFILER(FSocket::SendTo(Data,Count,BytesSent,Destination));
